login logic is refactored, need to do signup + homepage
