+++ lib/geocommunity/geocommunity.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 
-class CalendarWidget extends StatelessWidget {
-  const CalendarWidget({super.key});
+class GeocommunityWidget extends StatelessWidget {
+  const GeocommunityWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
     return const Text(
-      'THIS IS THE CALENDAR PAGE.',
+      'THIS IS THE GEOCOMMUNITY PAGE.',
       style: TextStyle(color: Colors.white, fontSize: 24)
     );
   }
