+++ lib/intro/intro.dart
@@ -2,7 +2,7 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import '../constants.dart';
-import '../home/home.dart';
+import '../nav.dart';
 import './signup.dart';
 import 'login.dart';
 
@@ -27,7 +27,7 @@ class _IntroState extends State<Intro> {
         if (user != null) {
           print("user is logged in! this is from intro.dart");
           print(user);
-          return const Home();
+          return const Nav();
         }
         
         print("user is NOT logged in!");
