+++ lib/leaderboard/leaderboard.dart
@@ -1,56 +1,13 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
 
-class Leaderboard extends StatefulWidget {
-  const Leaderboard({super.key});
-
-  @override
-  State<Leaderboard> createState() => _LeaderboardState();
-}
-
-class _LeaderboardState extends State<Leaderboard> {
-  int _selectedIndex = 2;
-  // static const TextStyle optionStyle = TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
-
-  final List<BottomNavigationBarItem> items = [
-    const BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
-    const BottomNavigationBarItem(
-        icon: Icon(Icons.calendar_today), label: 'Calendar'),
-    const BottomNavigationBarItem(
-        icon: Icon(Icons.leaderboard), label: 'Leaderboard'),
-    const BottomNavigationBarItem(icon: Icon(Icons.person), label: 'Profile'),
-  ];
-
-  void _onTabTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
+class LeaderboardWidget extends StatelessWidget {
+  const LeaderboardWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-            backgroundColor: Colors.transparent,
-            elevation: 0,
-            centerTitle: true,
-            // title: Image.asset('assets/logos/questify_1152.png',
-            //     width: 40, height: 40, fit: BoxFit.fitHeight)),
-        ),
-        body: Center(
-          child: widgetOptions.elementAt(_selectedIndex),
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          currentIndex: _selectedIndex,
-          type: BottomNavigationBarType.fixed,
-          onTap: _onTabTapped,
-          unselectedItemColor: const Color(0xFFE86500),
-          selectedItemColor: const Color(0xFFe69d65), // Color(0xFFf76c00),
-          showSelectedLabels: false,
-          showUnselectedLabels: false,
-          elevation: 0,
-          items: items,
-          backgroundColor: Colors.black,
-        ));
+    return const Text(
+      'THIS IS THE LEADERBOARD PAGE.',
+      style: TextStyle(color: Colors.white, fontSize: 24)
+    );
   }
 }
