--- lib/leaderboard/widgets/main_widget.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class LeaderboardWidget extends StatelessWidget {
-  const LeaderboardWidget({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const Text(
-      'THIS IS THE LEADERBOARD PAGE.',
-      style: TextStyle(color: Colors.white, fontSize: 24)
-    );
-  }
-}
