+++ lib/nav.dart
@@ -1,24 +1,24 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
+import 'constants.dart';
 
-class Calendar extends StatefulWidget {
-  const Calendar({super.key});
+class Nav extends StatefulWidget {
+  const Nav({super.key});
 
   @override
-  State<Calendar> createState() => _CalendarState();
+  State<Nav> createState() => _NavState();
 }
 
-class _CalendarState extends State<Calendar> {
-  int _selectedIndex = 1;
+class _NavState extends State<Nav> {
+  int _selectedIndex = 0;
   // static const TextStyle optionStyle = TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
 
   final List<BottomNavigationBarItem> items = [
-    const BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
+    const BottomNavigationBarItem(icon: Icon(Icons.home_rounded), label: 'Home'),
     const BottomNavigationBarItem(
-        icon: Icon(Icons.calendar_today), label: 'Calendar'),
+        icon: Icon(Icons.language_rounded), label: 'Geocommunity'),
     const BottomNavigationBarItem(
-        icon: Icon(Icons.leaderboard), label: 'Leaderboard'),
-    const BottomNavigationBarItem(icon: Icon(Icons.person), label: 'Profile'),
+        icon: Icon(Icons.arrow_outward_rounded), label: 'Stats'),
+    const BottomNavigationBarItem(icon: Icon(Icons.person_rounded), label: 'Profile'),
   ];
 
   void _onTabTapped(int index) {
@@ -34,23 +34,25 @@ class _CalendarState extends State<Calendar> {
             backgroundColor: Colors.transparent,
             elevation: 0,
             centerTitle: true,
+            // for the below line: https://stackoverflow.com/questions/44978216/flutter-remove-back-button-on-appbar
+            automaticallyImplyLeading: false,
             // title: Image.asset('assets/logos/questify_1152.png',
             //     width: 40, height: 40, fit: BoxFit.fitHeight)),
         ),
-        body: Center(
+        body: SingleChildScrollView(
           child: widgetOptions.elementAt(_selectedIndex),
         ),
         bottomNavigationBar: BottomNavigationBar(
           currentIndex: _selectedIndex,
           type: BottomNavigationBarType.fixed,
           onTap: _onTabTapped,
-          unselectedItemColor: const Color(0xFFE86500),
-          selectedItemColor: const Color(0xFFe69d65), // Color(0xFFf76c00),
+          unselectedItemColor: Colors.white.withOpacity(0.5),
+          selectedItemColor: const Color(0xFFFFFFFF), // Color(0xFFf76c00),
           showSelectedLabels: false,
           showUnselectedLabels: false,
           elevation: 0,
           items: items,
-          backgroundColor: Colors.black,
+          backgroundColor: const Color(0xFF282828),
         ));
   }
 }
