+++ lib/user_profile/user_profile.dart
@@ -1,57 +1,26 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 
-class UserProfile extends StatefulWidget {
-  const UserProfile({super.key});
+import '../../intro/intro.dart';
 
-  @override
-  State<UserProfile> createState() => _UserProfileState();
-}
-
-class _UserProfileState extends State<UserProfile> {
-  int _selectedIndex = 3;
-  // static const TextStyle optionStyle = TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
-
-  final List<BottomNavigationBarItem> items = [
-    const BottomNavigationBarItem(icon: Icon(Icons.home), label: 'Home'),
-    const BottomNavigationBarItem(
-        icon: Icon(Icons.calendar_today), label: 'Calendar'),
-    const BottomNavigationBarItem(icon: Icon(Icons.mic), label: 'Microphone'),
-    const BottomNavigationBarItem(
-        icon: Icon(Icons.leaderboard), label: 'Leaderboard'),
-    const BottomNavigationBarItem(icon: Icon(Icons.person), label: 'Profile'),
-  ];
-
-  void _onTabTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
-  }
+class UserProfileWidget extends StatelessWidget {
+  const UserProfileWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-        appBar: AppBar(
-            backgroundColor: Colors.transparent,
-            elevation: 0,
-            centerTitle: true,
-            // title: Image.asset('assets/logos/questify_1152.png',
-            //     width: 40, height: 40, fit: BoxFit.fitHeight)),
-        ),
-        body: Center(
-          child: widgetOptions.elementAt(_selectedIndex),
-        ),
-        bottomNavigationBar: BottomNavigationBar(
-          currentIndex: _selectedIndex,
-          type: BottomNavigationBarType.fixed,
-          onTap: _onTabTapped,
-          unselectedItemColor: const Color(0xFFE86500),
-          selectedItemColor: const Color(0xFFe69d65), // Color(0xFFf76c00),
-          showSelectedLabels: false,
-          showUnselectedLabels: false,
-          elevation: 0,
-          items: items,
-          backgroundColor: Colors.black,
-        ));
+    return TextButton(
+      onPressed: () {
+        FirebaseAuth.instance.signOut();
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => const Intro())
+        );
+        print('PRESSED SIGNED OUT');
+      },
+      child: Text(
+        'Sign out',
+        style: TextStyle(color: Colors.white, fontSize: 24)
+      )
+    );
   }
 }
