+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   cloud_firestore: ^4.8.1
   table_calendar: ^3.0.9
   firebase_auth: ^4.6.3
+  shared_preferences: ^2.2.0
 
 dev_dependencies:
   flutter_test:
